refactor(llm/glm): move inline tests to glm/tests.rs
Move #[cfg(test)] mod tests from glm.rs to glm/tests.rs module file.
glm.rs -> glm/mod.rs (241 lines), tests extracted to glm/tests.rs (190 lines).

Source: issue #379
Type: refactor
